Move decodeError to the abi module.
