+++ packages/core/src/new-api/internal/new-execution/abi.ts
@@ -13,11 +13,29 @@ import {
   EvmExecutionResultTypes,
   EvmTuple,
   EvmValue,
+  FailedEvmExecutionResult,
   InvalidResultError,
   RevertWithCustomError,
   RevertWithInvalidData,
+  RevertWithPanicCode,
+  RevertWithReason,
   SuccessfulEvmExecutionResult,
 } from "./types/evm-execution";
+
+const REVERT_REASON_SIGNATURE = "0x08c379a0";
+const PANIC_CODE_SIGNATURE = "0x4e487b71";
+const PANIC_CODE_NAMES: { [key: number]: string | undefined } = {
+  [0x00]: "GENERIC_PANIC",
+  [0x01]: "ASSERT_FALSE",
+  [0x11]: "OVERFLOW",
+  [0x12]: "DIVIDE_BY_ZERO",
+  [0x21]: "ENUM_RANGE_ERROR",
+  [0x22]: "BAD_STORAGE_DATA",
+  [0x31]: "STACK_UNDERFLOW",
+  [0x32]: "ARRAY_RANGE_ERROR",
+  [0x41]: "OUT_OF_MEMORY",
+  [0x51]: "UNINITIALIZED_FUNCTION_CALL",
+};
 
 /**
  * Links the libraries in the artifact's deployment bytecode, encodes the constructor
@@ -117,38 +135,6 @@ export function decodeArtifactFunctionCallResult(
 }
 
 /**
- * Returns a function fragment for the given function name in the given artifact.
- *
- * @param artifact The artifact to search in.
- * @param functionName The function name to search for. MUST be validated first.
- */
-function getFunctionFragment(
-  iface: Interface,
-  functionName: string
-): FunctionFragment {
-  const { ethers } = require("ethers") as typeof import("ethers");
-
-  const fragment = iface.fragments
-    .filter(ethers.Fragment.isFunction)
-    .find(
-      (fr) =>
-        fr.name === functionName ||
-        getFunctionNameWithParams(fr) === functionName
-    );
-
-  assertIgnitionInvariant(
-    fragment !== undefined,
-    "Called getFunctionFragment with an invalid function name"
-  );
-
-  return fragment;
-}
-
-function getFunctionNameWithParams(functionFragment: FunctionFragment): string {
-  return functionFragment.format("sighash");
-}
-
-/**
  * Validates that a function name is valid for the given artifact. That means:
  *  - It's a valid function name
  *  - The function name exists in the artifact's ABI
@@ -219,6 +205,129 @@ ${normalizedFunctionNameList}`
 
 ${normalizedFunctionNameList}`
     );
+  }
+}
+
+/**
+ * Decodes an error from a failed evm execution.
+ *
+ * @param returnData The data, as returned by the JSON-RPC.
+ * @param customErrorReported A value indicating if the JSON-RPC error
+ *  reported that it was due to a custom error.
+ * @param decodeCustomError A function that decodes custom errors, returning
+ *  `RevertWithCustomError` if succesfully decoded, `RevertWithInvalidData`
+ *  if a custom error was recognized but couldn't be decoded, and `undefined`
+ *  it it wasn't recognized.
+ * @returns A `FailedEvmExecutionResult` with the decoded error.
+ */
+export function decodeError(
+  returnData: string,
+  customErrorReported: boolean,
+  decodeCustomError?: (
+    returnData: string
+  ) => RevertWithCustomError | RevertWithInvalidData | undefined
+): FailedEvmExecutionResult {
+  if (returnData === "0x") {
+    return { type: EvmExecutionResultTypes.REVERT_WITHOUT_REASON };
+  }
+
+  const revertWithReasonError = tryToDecodeReason(returnData);
+  if (revertWithReasonError !== undefined) {
+    return revertWithReasonError;
+  }
+
+  const revertWithPanicCodeError = tryToDecodePanicCode(returnData);
+  if (revertWithPanicCodeError !== undefined) {
+    return revertWithPanicCodeError;
+  }
+
+  if (decodeCustomError !== undefined) {
+    const decodedCustomError = decodeCustomError(returnData);
+    if (decodedCustomError !== undefined) {
+      return decodedCustomError;
+    }
+  }
+
+  if (customErrorReported === true) {
+    return {
+      type: EvmExecutionResultTypes.REVERT_WITH_UNKNOWN_CUSTOM_ERROR,
+      signature: returnData.slice(0, 10),
+      data: returnData,
+    };
+  }
+
+  return {
+    type: EvmExecutionResultTypes.REVERT_WITH_INVALID_DATA_OR_UNKNOWN_CUSTOM_ERROR,
+    signature: returnData.slice(0, 10),
+    data: returnData,
+  };
+}
+
+function tryToDecodeReason(
+  returnData: string
+): RevertWithReason | RevertWithInvalidData | undefined {
+  if (!returnData.startsWith(REVERT_REASON_SIGNATURE)) {
+    return undefined;
+  }
+
+  const { ethers } = require("ethers") as typeof import("ethers");
+  const abiCoder = ethers.AbiCoder.defaultAbiCoder();
+
+  try {
+    const [reason] = abiCoder.decode(
+      ["string"],
+      Buffer.from(returnData.slice(REVERT_REASON_SIGNATURE.length), "hex")
+    );
+
+    return {
+      type: EvmExecutionResultTypes.REVERT_WITH_REASON,
+      message: reason,
+    };
+  } catch {
+    return {
+      type: EvmExecutionResultTypes.REVERT_WITH_INVALID_DATA,
+      data: returnData,
+    };
+  }
+}
+
+function tryToDecodePanicCode(
+  returnData: string
+): RevertWithPanicCode | RevertWithInvalidData | undefined {
+  if (!returnData.startsWith(PANIC_CODE_SIGNATURE)) {
+    return undefined;
+  }
+
+  const { ethers } = require("ethers") as typeof import("ethers");
+  const abiCoder = ethers.AbiCoder.defaultAbiCoder();
+
+  try {
+    const decoded = abiCoder.decode(
+      ["uint256"],
+      Buffer.from(returnData.slice(REVERT_REASON_SIGNATURE.length), "hex")
+    );
+
+    const panicCode = Number(decoded[0]);
+
+    const panicName = PANIC_CODE_NAMES[panicCode];
+
+    if (panicName === undefined) {
+      return {
+        type: EvmExecutionResultTypes.REVERT_WITH_INVALID_DATA,
+        data: returnData,
+      };
+    }
+
+    return {
+      type: EvmExecutionResultTypes.REVERT_WITH_PANIC_CODE,
+      panicCode,
+      panicName,
+    };
+  } catch {
+    return {
+      type: EvmExecutionResultTypes.REVERT_WITH_INVALID_DATA,
+      data: returnData,
+    };
   }
 }
 
@@ -299,4 +408,36 @@ function ethersResultIntoEvmTuple(
   }
 
   return { positional, named };
+}
+
+/**
+ * Returns a function fragment for the given function name in the given artifact.
+ *
+ * @param artifact The artifact to search in.
+ * @param functionName The function name to search for. MUST be validated first.
+ */
+function getFunctionFragment(
+  iface: Interface,
+  functionName: string
+): FunctionFragment {
+  const { ethers } = require("ethers") as typeof import("ethers");
+
+  const fragment = iface.fragments
+    .filter(ethers.Fragment.isFunction)
+    .find(
+      (fr) =>
+        fr.name === functionName ||
+        getFunctionNameWithParams(fr) === functionName
+    );
+
+  assertIgnitionInvariant(
+    fragment !== undefined,
+    "Called getFunctionFragment with an invalid function name"
+  );
+
+  return fragment;
+}
+
+function getFunctionNameWithParams(functionFragment: FunctionFragment): string {
+  return functionFragment.format("sighash");
 }
